+++ src/main/kotlin/bridge/model/BridgeMaker.kt
@@ -1,4 +1,6 @@
-package bridge
+package bridge.model
+
+import bridge.BridgeNumberGenerator
 
 /**
  * 다리의 길이를 입력 받아서 다리를 생성해주는 역할을 한다.
