+++ src/test/kotlin/bridge/ApplicationTest.kt
@@ -1,5 +1,6 @@
 package bridge
 
+import bridge.model.BridgeMaker
 import camp.nextstep.edu.missionutils.test.Assertions.assertRandomNumberInRangeTest
 import camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest
 import camp.nextstep.edu.missionutils.test.NsTest
